feat: show hoover graphic on about
